# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API quản lý OpenVPN, Authentication, Portal...",
        "title": "System Portal API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/portal/permissions": {
            "get": {
                "security": [
                    {"BearerAuth": []}
                ],
                "produces": ["application/json"],
                "tags": ["Permissions"],
                "summary": "List permissions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {"$ref": "#/definitions/entities.Permission"}
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {"BearerAuth": []}
                ],
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Permissions"],
                "summary": "Create permission",
                "parameters": [
                    {
                        "description": "Permission data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {"$ref": "#/definitions/entities.Permission"}
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {"$ref": "#/definitions/response.SuccessResponse"}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {"$ref": "#/definitions/response.ErrorResponse"}
                    }
                }
            }
        },
        "/api/portal/permissions/{id}": {
            "put": {
                "security": [
                    {"BearerAuth": []}
                ],
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Permissions"],
                "summary": "Update permission",
                "parameters": [
                    {"type": "string", "description": "Permission ID", "name": "id", "in": "path", "required": true},
                    {"description": "Permission data", "name": "request", "in": "body", "required": true, "schema": {"$ref": "#/definitions/entities.Permission"}}
                ],
                "responses": {
                    "200": {"description": "OK", "schema": {"$ref": "#/definitions/response.SuccessResponse"}},
                    "400": {"description": "Bad Request", "schema": {"$ref": "#/definitions/response.ErrorResponse"}}
                }
            },
            "delete": {
                "security": [
                    {"BearerAuth": []}
                ],
                "produces": ["application/json"],
                "tags": ["Permissions"],
                "summary": "Delete permission",
                "parameters": [
                    {"type": "string", "description": "Permission ID", "name": "id", "in": "path", "required": true}
                ],
                "responses": {
                    "200": {"description": "OK", "schema": {"$ref": "#/definitions/response.SuccessResponse"}},
                    "400": {"description": "Bad Request", "schema": {"$ref": "#/definitions/response.ErrorResponse"}}
                }
            }
        },
        "/api/portal/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all portal users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Users"
                ],
                "summary": "List portal users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PortalUserResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new portal user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Users"
                ],
                "summary": "Create portal user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PortalUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/portal/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get portal user by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Users"
                ],
                "summary": "Get portal user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PortalUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update information for a portal user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Users"
                ],
                "summary": "Update portal user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PortalUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a portal user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Users"
                ],
                "summary": "Delete portal user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/portal/users/{id}/activate": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Users"
                ],
                "summary": "Activate portal user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/portal/users/{id}/deactivate": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Users"
                ],
                "summary": "Deactivate portal user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/portal/users/{id}/reset-password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Users"
                ],
                "summary": "Reset portal user password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.PortalUserRequest": {
            "type": "object",
            "required": [
                "email",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "groupId": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.PortalUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "groupId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "entities.Permission": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "details": {},
                        "message": {
                            "type": "string"
                        },
                        "status": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "object",
                    "properties": {
                        "data": {},
                        "message": {
                            "type": "string"
                        },
                        "status": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}